tier_synthesis/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o tier_synthesis/requirements.txt
anyio==4.8.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
apsw==3.48.0.0
    # via apswutils
apswutils==0.0.2
    # via fastlite
beautifulsoup4==4.12.3
    # via python-fasthtml
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
click==8.1.8
    # via uvicorn
fastcore==1.8.14
    # via
    #   apswutils
    #   fastlite
    #   python-fasthtml
fastlite==0.1.1
    # via python-fasthtml
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via python-fasthtml
idna==3.10
    # via
    #   anyio
    #   httpx
itsdangerous==2.2.0
    # via python-fasthtml
joblib==1.5.2
    # via scikit-learn
numpy==2.3.4
    # via
    #   tier-synthesis (pyproject.toml)
    #   scikit-learn
    #   scipy
oauthlib==3.2.2
    # via python-fasthtml
packaging==24.2
    # via fastcore
pillow==11.1.0
    # via tier-synthesis (pyproject.toml)
python-dateutil==2.9.0.post0
    # via python-fasthtml
python-dotenv==1.0.1
    # via uvicorn
python-fasthtml==0.12.33
    # via tier-synthesis (pyproject.toml)
python-multipart==0.0.20
    # via python-fasthtml
pyyaml==6.0.2
    # via uvicorn
scikit-learn==1.7.2
    # via tier-synthesis (pyproject.toml)
scipy==1.16.3
    # via scikit-learn
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
soupsieve==2.6
    # via beautifulsoup4
starlette==0.45.2
    # via python-fasthtml
threadpoolctl==3.6.0
    # via scikit-learn
typing-extensions==4.15.0
    # via anyio
uvicorn==0.34.0
    # via python-fasthtml
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.4
    # via uvicorn
websockets==14.1
    # via uvicorn
